requirements.txt
flask~=3.1.1
flasgger~=0.9.7.1
requests~=2.32.4
beautifulsoup4~=4.13.4
python-dotenv~=1.1.1
crewai~=0.152.0
langchain
langchain-openai~=0.3.28
flasgger

certifi~=2025.7.14
streamlit~=1.47.1
protobuf~=5.29.5
openai~=1.98.0